requirements.txt
astroid==2.1.0
autopep8==1.4.3
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
curlify==2.1.1
decorator==4.3.2
Django==2.1.5
django-environ==0.4.5
django-extensions==2.1.5
django-rest-framework==0.1.0
djangorestframework==3.9.1
facebook-business==3.2.6
facebookads==2.11.4
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.3.2
pickleshare==0.7.5
prompt-toolkit==2.0.8
pycodestyle==2.5.0
Pygments==2.3.1
pylint==2.2.2
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
wrapt==1.11.1
